refactor(core): split facet wrap/grid and point/path geometry
Extract facet helpers, wrap partition, and grid partition behind resolveFacet;
split points vs line/area path builders out of geometry-point-path. Tests pin
empty grid combos, form errors, and closed area batches.
