+++ packages/core/src/pipeline/geometry-points.ts
@@ -1,0 +1,65 @@
+/**
+ * Point mark geometry batch builder.
+ */
+import type { PointsBatch } from "../scene.js";
+
+import type { LayerFrame, PipelineWarning, ResolvedColorScale } from "./types.js";
+import { colorOf } from "./types.js";
+import type { Frame } from "./geometry-shared.js";
+import { DEFAULT_POINT_SIZE, positionOf, removedWarning } from "./geometry-shared.js";
+
+export function pointsBatch(
+  frame: LayerFrame,
+  fx: Frame,
+  color: ResolvedColorScale | null,
+  warnings: PipelineWarning[],
+): PointsBatch | null {
+  const { binding, n } = frame;
+  const xs = new Float64Array(n);
+  const ys = new Float64Array(n);
+  const keptRows = new Uint32Array(n);
+  let kept = 0;
+  for (let row = 0; row < n; row++) {
+    const tx = positionOf(fx.xScale, frame.xNumeric, frame.xValues, row, frame.offsetX);
+    const ty = positionOf(fx.yScale, frame.yNumeric, null, row, frame.offsetY);
+    if (Number.isNaN(tx) || Number.isNaN(ty)) continue;
+    xs[kept] = tx;
+    ys[kept] = ty;
+    keptRows[kept] = row;
+    kept++;
+  }
+  removedWarning(n - kept, binding.index, warnings);
+  if (kept === 0) return null;
+
+  const positions = new Float32Array(kept * 2);
+  const rowIndex = new Uint32Array(kept);
+  for (let j = 0; j < kept; j++) {
+    positions[j * 2] = xs[j]! * fx.innerWidth;
+    positions[j * 2 + 1] = fx.innerHeight - ys[j]! * fx.innerHeight;
+    rowIndex[j] = frame.rowIndex[keptRows[j]!]!;
+  }
+
+  const params = binding.layer.geom === "point" ? (binding.layer.params ?? {}) : {};
+  const batch: PointsBatch = {
+    kind: "points",
+    layerIndex: binding.index,
+    panelIndex: 0,
+    positions,
+    rowIndex,
+    size: params.size ?? DEFAULT_POINT_SIZE,
+    alpha: params.alpha ?? 1,
+    shape: params.shape ?? "circle",
+    fill: binding.color.constant,
+  };
+  if (color !== null && (frame.colorValues !== null || binding.color.scaledConstant !== null)) {
+    const colors = Array.from<string>({ length: kept });
+    for (let j = 0; j < kept; j++) {
+      const row = keptRows[j]!;
+      const value =
+        frame.colorValues === null ? binding.color.scaledConstant! : frame.colorValues[row]!;
+      colors[j] = colorOf(color, value);
+    }
+    batch.colors = colors;
+  }
+  return batch;
+}
